refactor:implement modular domain architecture for better layout (core, vanguard, recovery, intelligence)
