+++ src/sentinel/core/hooks.py
