+++ src/sentinel/intelligence/diagnose.py
