+++ src/sentinel/vanguard/boot.py
